Code divided in widget for a better readability
